+++ frontend/src/hooks/useIsGithubLoginValid/index.tsx
@@ -1,8 +1,8 @@
 import { gql } from "@apollo/client";
 import { useHasuraLazyQuery } from "src/hooks/useHasuraQuery";
-import { GITHUB_USER_FRAGMENT } from "src/pages/ProjectDetails/PaymentActions/useGetPaymentRequests";
 import { HasuraUserRole } from "src/types";
 import { FindUserQueryForPaymentFormQuery } from "src/__generated/graphql";
+import { GITHUB_USER_FRAGMENT } from "../usePaymentRequests";
 
 export default function useFindGithubUser() {
   const [trigger, query] = useHasuraLazyQuery<FindUserQueryForPaymentFormQuery>(
