+++ frontend/src/pages/ProjectDetails/PaymentActions/PaymentActions.test.tsx
@@ -7,7 +7,7 @@ import { RoutePaths } from "src/App";
 import { LOCAL_STORAGE_TOKEN_SET_KEY } from "src/hooks/useTokenSet";
 import { ProjectDetailsProvider } from "../ProjectDetailsContext";
 import { GithubUserFragment, PaymentRequestFragment } from "src/__generated/graphql";
-import { GET_GITHUB_USER_QUERY, PAYMENT_REQUESTS_FOR_PROJECT_SUBSCRIPTION } from "./useGetPaymentRequests";
+import { GET_GITHUB_USER_QUERY, PAYMENT_REQUESTS_FOR_PROJECT_SUBSCRIPTION } from "src/hooks/usePaymentRequests";
 
 expect.extend(matchers);
