+++ frontend/src/pages/ProjectDetails/PaymentActions/index.tsx
@@ -4,6 +4,7 @@ import PaymentTable from "src/components/PaymentTable";
 import ProjectPaymentTableFallback from "src/components/ProjectPaymentTableFallback";
 import QueryWrapper from "src/components/QueryWrapper";
 import { useIntl } from "src/hooks/useIntl";
+import usePaymentRequests from "src/hooks/usePaymentRequests";
 import {
   PaymentAction,
   ProjectDetailsActionType,
@@ -12,7 +13,6 @@ import {
 } from "../ProjectDetailsContext";
 import PaymentForm from "./PaymentForm";
 import RemainingBudget from "./RemainingBudget";
-import useGetPaymentRequests from "./useGetPaymentRequests";
 
 interface PaymentsProps {
   projectId: string;
@@ -24,7 +24,7 @@ export default function PaymentActions({ projectId }: PaymentsProps) {
   const state = useContext(ProjectDetailsContext);
   const dispatch = useContext(ProjectDetailsDispatchContext);
 
-  const query = useGetPaymentRequests(projectId);
+  const query = usePaymentRequests(projectId);
   const payments = query.data?.paymentRequests || [];
   const budget = query.data?.budget || { initialAmount: 0, remainingAmount: 0 };
